Move solver pro app to its own file
This was done since `type(SolverProApp)` would be different when importing
it from another module/file. In the main execution file, `SolverProApp`
is declared as a type under `__main__`, which is not the same as the module
type `src.app.mainApp.SolverProApp` (which seems like a bug to me but whatever).