Revert tests to use original contradicting relations
This was done because, technically speaking, the newest relation is the
one that's causing the contradiction. Even if it is a restriction case,
if it is a bad restriction (ie restricts against some previous relation
defined on expression lists), it is a contradiction on the part of the
relation trying to restrict the variable, not the originally defining
relation.