Fix restrict redef checking
If you solve for a symbol `c` which has conditions based on `a` and `b`,
then you try to restrict `b`, this algorithm removes `b` from the
database. Then later, when the substituter tries to check the conditions
on `c`, it results in a KeyError, because `b` isn't a "known variable"
anymore. The fix: Tell the substituter about `b`, and if it runs into an
unknown symbol in a condition, it asserts that symbol is the one
potentially being restricted.